backend/requirements/base.txt
Django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
django-filter==24.1
psycopg2-binary==2.9.*
Pillow==10.*
WeasyPrint==61.*
gunicorn==21.*
python-decouple==3.8
dj-database-url==2.1.*
django-storages[s3]==1.14.*
boto3==1.34.*
channels==4.0.*
channels-redis==4.2.*
daphne==4.1.*
